+++ src/tree.rs
@@ -3,11 +3,11 @@ fixed id (0 or -1), and its parent attribute points at the third child.
     The distance for the tree root will be set to -inf.
 */
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct TreeNodeId<Id> {
     pub parent: Id,
     pub left: Option<Id>,
     pub right: Option<Id>,
 }
 
-pub type TreeNode = TreeNodeId<usize>;
+pub type TreeNode = TreeNodeId<u32>;
